Add a heroku log device using logfmt
