+++ lib/timber/contexts/http_request.rb
@@ -6,7 +6,7 @@ module Timber
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :headers => headers,
               :host => host,
