+++ lib/timber/contexts/http_requests/action_controller_specific.rb
@@ -30,15 +30,15 @@ module Timber
 
         private
           def json_payload
-            @json_payload ||= DeepMerger.merge(super, {
-              Rack._root_key => {
+            @json_payload ||= {
+              Rack._root_key => Core::DeepMerger.merge(super, {
                 _root_key => {
                   :action => action,
                   :controller => controller,
                   :format => format
                 }
-              }
-            })
+              })
+            }
           end
       end
     end
