+++ lib/timber/contexts/http_response.rb
@@ -7,7 +7,7 @@ module Timber
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :headers => headers,
               :status => status,
