+++ lib/timber/contexts/logger.rb
@@ -14,7 +14,7 @@ module Timber
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :level => level,
               :progname => progname
