+++ lib/timber/contexts/server.rb
@@ -7,12 +7,12 @@ module Timber
       VERSION = 1.freeze
 
       def hostname
-        @hostname ||= Socket.gethostname
+        @hostname ||= ::Socket.gethostname
       end
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :hostname => hostname
             }
