+++ lib/timber/contexts/sql_queries/active_record_specific.rb
@@ -37,16 +37,16 @@ module Timber
 
         private
           def json_payload
-            @json_payload ||= DeepMerger.merge(super, {
-              SQLQuery._root_key => {
+            @json_payload ||= {
+              SQLQuery._root_key => Core::DeepMerger.merge(super, {
                 _root_key => {
                   :binds => binds,
                   :connection_id => connection_id,
                   :statement_name => statement_name,
                   :transaction_id => transaction_id
                 }
-              }
-            })
+              })
+            }
           end
 
           def payload
