+++ lib/timber/contexts/sql_query.rb
@@ -7,7 +7,7 @@ module Timber
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :sql => sql,
               :time_ms => time_ms
