+++ lib/timber/contexts/template_render.rb
@@ -6,7 +6,7 @@ module Timber
 
       private
         def json_payload
-          @json_payload ||= DeepMerger.merge(super, {
+          @json_payload ||= Core::DeepMerger.merge(super, {
             _root_key => {
               :name => name,
               :time_ms => time_ms
