+++ lib/timber/core.rb
@@ -1,0 +1,2 @@
+require "timber/core/deep_merger"
+require "timber/core/logfmt_encoder"
