+++ lib/timber/core/deep_merger.rb
@@ -1,0 +1,11 @@
+module Timber
+  module Core
+    # Deep merges hash keys
+    module DeepMerger
+      def self.merge(first, second)
+        merger = proc { |key, v1, v2| Hash === v1 && Hash === v2 ? v1.merge(v2, &merger) : v2 }
+        first.merge(second, &merger)
+      end
+    end
+  end
+end
