+++ lib/timber/log_device.rb
@@ -1,0 +1,5 @@
+module Timber
+  # Placeholder base class for all Timber log devices
+  class LogDevice
+  end
+end
