+++ lib/timber/log_devices.rb
@@ -1,0 +1,4 @@
+# Order is relevant
+require File.join(File.dirname(__FILE__), "log_devices", "io")
+require File.join(File.dirname(__FILE__), "log_devices", "heroku_logplex")
+require File.join(File.dirname(__FILE__), "log_devices", "http")
