+++ lib/timber/patterns.rb
@@ -1,2 +1,3 @@
 require "timber/patterns/delegated_singleton"
 require "timber/patterns/to_json"
+require "timber/patterns/to_logfmt"
