+++ spec/timber/probes/action_controller_log_subscriber/action_controller_spec.rb
@@ -7,7 +7,6 @@ describe Timber::Probes::ActiveSupportLogSubscriber::ActionController do
 
       def index
         response.headers['Content-Length'] = "500"
-        raise Timber::CurrentContext.snapshot.to_json.inspect
         render json: {}
       end
 
@@ -25,11 +24,11 @@ describe Timber::Probes::ActiveSupportLogSubscriber::ActionController do
     Object.send(:remove_const, :PagesController)
   end
 
-  let(:response_context_class) { Timber::Contexts::ActionControllerResponse }
+  let(:response_context_class) { Timber::Contexts::HTTPResponses::ActionController }
 
   describe "#process_action" do
     it "should set the context" do
-      #expect_any_instance_of(response_context_class).to receive(:event=).with(kind_of(ActiveSupport::Notifications::Event)).once
+      expect_any_instance_of(response_context_class).to receive(:event=).with(kind_of(ActiveSupport::Notifications::Event)).once
       dispatch_rails_request("/pages")
     end
   end
